Fix debug mode startup errors on greedy voxels.

+++ src/modes/playgame/vox/greedy.rs
@@ -1,7 +1,7 @@
 use super::modelsize::ModelSize;
 use std::collections::{HashMap, HashSet};
 
-pub type VoxMap = HashMap<u32, u8>;
+pub type VoxMap = HashMap<i32, u8>;
 
 pub fn greedy_cubes(cube_index: &mut VoxMap, output: &mut Vec<f32>, size: &ModelSize) {
     let invisible = cube_index
@@ -9,13 +9,13 @@ pub fn greedy_cubes(cube_index: &mut VoxMap, output: &mut Vec<f32>, size: &Model
         .filter(|(idx, _)| {
             cube_index.contains_key(&(*idx - 1))
                 && cube_index.contains_key(&(*idx + 1))
-                && cube_index.contains_key(&(*idx - size.x))
-                && cube_index.contains_key(&(*idx + size.x))
-                && cube_index.contains_key(&(*idx + (size.x * size.x)))
-                && cube_index.contains_key(&(*idx - (size.x * size.x)))
+                && cube_index.contains_key(&(*idx - size.x as i32))
+                && cube_index.contains_key(&(*idx + size.x as i32))
+                && cube_index.contains_key(&(*idx + (size.x * size.x) as i32))
+                && cube_index.contains_key(&(*idx - (size.x * size.x) as i32))
         })
         .map(|(idx, _)| *idx)
-        .collect::<HashSet<u32>>();
+        .collect::<HashSet<i32>>();
     //println!("Invisibility cull: {}", invisible.len());
     cube_index.retain(|idx, _| !invisible.contains(idx));
 
@@ -27,9 +27,9 @@ pub fn greedy_cubes(cube_index: &mut VoxMap, output: &mut Vec<f32>, size: &Model
             let idx = *min_iter.unwrap();
             let mat_idx = cube_index.remove(&idx).unwrap();
 
-            let (x, y, z) = size.idxmap(idx);
-            let width = grow_right(cube_index, idx, mat_idx);
-            let height = grow_down(cube_index, idx, width, mat_idx, size);
+            let (x, y, z) = size.idxmap(idx as u32);
+            let width = grow_right(cube_index, idx as u32, mat_idx);
+            let height = grow_down(cube_index, idx as u32, width, mat_idx, size);
             //let depth = grow_in(&mut cube_index, idx, width, height);
             let depth = 1;
 
@@ -51,8 +51,8 @@ fn grow_right(cube_index: &mut VoxMap, idx: u32, mat: u8) -> u32 {
     let mut width = 1;
     let mut candidate_idx = idx + 1;
 
-    while cube_index.contains_key(&candidate_idx) && cube_index[&candidate_idx] == mat {
-        cube_index.remove(&candidate_idx);
+    while cube_index.contains_key(&(candidate_idx as i32)) && cube_index[&(candidate_idx as i32)] == mat {
+        cube_index.remove(&(candidate_idx as i32));
         width += 1;
         candidate_idx += 1;
     }
@@ -65,16 +65,16 @@ fn grow_down(cube_index: &mut VoxMap, idx: u32, width: u32, mat: u8, size: &Mode
     let mut candidate_idx = idx + size.x;
     'outer: loop {
         for cidx in candidate_idx..candidate_idx + width {
-            if !cube_index.contains_key(&cidx) {
+            if !cube_index.contains_key(&(cidx as i32)) {
                 break 'outer;
             }
-            if cube_index[&cidx] != mat {
+            if cube_index[&(cidx as i32)] != mat {
                 break 'outer;
             }
         }
 
         for cidx in candidate_idx..candidate_idx + width {
-            cube_index.remove(&cidx);
+            cube_index.remove(&(cidx as i32));
         }
         height += 1;
         candidate_idx += size.x;
